create m3u file from playlist, fixes to playlist endpoints

+++ src/Data/Repositories/PlaylistRepository.cs
@@ -23,14 +23,25 @@ public class PlaylistRepository(IDbContext dbContext) : IPlaylistRepository
         return (await conn.QueryAsync<Playlist, PlaylistEntry, Video, Playlist>(@"
         select *
         from playlist p
-        join playlist_entry pe on p.id = pe.playlist_id
-        join video v on pe.video_id = v.id 
+        left join playlist_entry pe on p.id = pe.playlistId
+        left join video v on pe.videoId = v.id 
         where p.id = @id", (playlist, playlistEntry, video) => {
-            playlistEntry.video = video;
-            playlist.entries.Add(playlistEntry);
+            if(playlistEntry != null) {
+                playlistEntry.video = video;
+                playlist.entries.Add(playlistEntry);
+            }
             return playlist;
-        }, new { id }))?.FirstOrDefault();
+        }, new { id }, splitOn: "id, id"))
+            .GroupBy(p => p.id)
+            .Select((grp) => {
+                var p = grp.First();
+                p.entries = grp.Where(p => p.entries.Count != 0).Select(p => p.entries.Single()).ToList();
+                return p;
+            })
+            .FirstOrDefault();
     }
+
+
 
     public async Task<Playlist> CreateAsync(PlaylistCreateRequest request)
     {
@@ -44,33 +55,40 @@ public class PlaylistRepository(IDbContext dbContext) : IPlaylistRepository
 
     public async Task<Playlist?> UpdateEntriesAsync(int playlistId, PlaylistEntryUpdateRequest request) {
         using var conn = dbContext.DbConnection;
+        conn.Open();
         using var transaction = conn.BeginTransaction();
 
         try {
 
+        var tasks = new List<Task>();
 
-        var inserted = conn.ExecuteAsync(@"
-            INSERT INTO playlist_entry(playlistId, videoId, order)
-            VALUES (@playlistId, @videoId, @order)
-        ", request.toAdd, transaction: transaction);
+        if(request?.toAdd?.Count > 0) {
+            tasks.Add(conn.ExecuteAsync(@"
+            INSERT INTO playlist_entry(playlistId, videoId, entryorder)
+            VALUES (@playlistId, @videoId, @entryorder)
+        ", request.toAdd, transaction: transaction));
+        }
 
-        var deleted = conn.ExecuteAsync(@"
+        if(request?.toRemove?.Count > 0) {
+            tasks.Add(conn.ExecuteAsync(@"
             DELETE FROM playlist_entry WHERE id = @id
-        ", request.toRemove.Select(id => new { id }), transaction: transaction);
+        ", request.toRemove.Select(id => new { id }), transaction: transaction));
+        }
 
-        var updated = conn.ExecuteAsync(@"
-            UPDATE playlist_entry SET videoId = @videoId, order = @order
+        if(request?.toUpdate?.Count > 0) {
+            tasks.Add(conn.ExecuteAsync(@"
+            UPDATE playlist_entry SET videoId = @videoId, entryorder = @entryorder
             WHERE id = @id
-        ", request.toUpdate, transaction: transaction);
+        ", request.toUpdate, transaction: transaction));
+        }
 
-        await Task.WhenAll(inserted, deleted, updated);
+        await Task.WhenAll(tasks);
         transaction.Commit();
-
-        return await GetByIdAsync(playlistId);
-
         } catch {
             transaction.Rollback();
             throw;
         }
+
+        return await GetByIdAsync(playlistId);
     }
 }
